move streamIsValid to webrtc proxy
