+++ packages/common/src/webrtc/Peer.ts
@@ -1,7 +1,7 @@
 import logger from '../util/logger'
-import { getUserMedia, getMediaConstraints, streamIsValid, sdpStereoHack } from './helpers'
+import { getUserMedia, getMediaConstraints, sdpStereoHack } from './helpers'
 import { PeerType, SwEvent } from '../util/constants'
-import { attachMediaStream, sdpToJsonHack, RTCPeerConnection } from '../util/webrtc'
+import { attachMediaStream, sdpToJsonHack, RTCPeerConnection, streamIsValid } from '../util/webrtc'
 import { isFunction } from '../util/helpers'
 import { DialogOptions } from '../util/interfaces'
 import { trigger } from '../services/Handler'
